requirements.txt
streamlit
torch==2.2.0
torchvision==0.17.0
pillow
gdown
numpy
scikit-learn